Generate functions to define defaults (incomplete)
These functions will be run when the script calling auto-docopts
evaluates its output. This will be too late to influence the
output of docopts.

Because docopts sets default values for all options, it would make no
use to run these funtions before evaluating the docopts output.

Because docopts does not leave option variables unset when they
are not specified on the command line, it is impossible to discern
an ommitted option from one that is specified to the same value
as the default.

Thus defaults cannot be applied after the docopts output is evaluated.

Therefore these functions should be executed not at eval time,
but when docopts is running, and they should affect the input
to docopts.

Therefore, these functions should be in gawk, or whatever is
parsing, and altering, the docopts help string.